requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==4.0.7
django-extensions==3.0.2
djangorestframework==3.13.1
future==0.18.2
graphviz==0.14
gunicorn==20.1.0
html5lib==1.1
idna==3.2
mysql-connector-python==8.0.20
mysqlclient==2.0.3
Pillow==8.3.1
protobuf==3.12.2
pydot==1.4.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.1
reportlab==3.6.1
requests==2.26.0
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.5
